+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Calendar Versioning](https://calver.org/).
 
 - Repo interaction and user detail cleaning when removing a repo with existing downloaded data.
 - Warning on category assignments referring to IDs not defined in the tool list.
+- Active maintainer number.
 - Add the `get_scores.py` script to process the OpenSSF scores and a dedicated visualisation page in the dashboard.
 
 ### Fixed